Implement login and register accounts

+++ backend/src/models/applicationModel.js
@@ -1,7 +1,6 @@
 const pool = require("../database/database");
 
-async function createApplication(application) {
-
+async function createApplication(application, userId) {
     const {
         company,
         role,
@@ -14,6 +13,7 @@ async function createApplication(application) {
 
     const sql = `
         INSERT INTO applications (
+            user_id,
             company,
             role,
             location,
@@ -22,10 +22,11 @@ async function createApplication(application) {
             deadline,
             notes
         )
-        VALUES ($1, $2, $3, $4, $5, $6, $7)
+        VALUES ($1, $2, $3, $4, $5, $6, $7, $8)
     `;
 
     await pool.query(sql, [
+        userId,
         company,
         role,
         location,
@@ -36,41 +37,43 @@ async function createApplication(application) {
     ]);
 }
 
-async function getApplications() {
-
+async function getApplications(userId) {
     const sql = `
         SELECT *
         FROM applications
+        WHERE user_id = $1
+        ORDER BY created_at DESC
     `;
 
-    const result = await pool.query(sql);
+    const result = await pool.query(sql, [userId]);
 
     return result.rows;
 }
 
-async function updateStatus(id, newStatus) {
-
+async function updateStatus(id, newStatus, userId) {
     const sql = `
         UPDATE applications
-        SET status = $1
+        SET status = $1,
+            updated_at = CURRENT_TIMESTAMP
         WHERE id = $2
+        AND user_id = $3
     `;
 
-    await pool.query(sql, [newStatus, id]);
+    await pool.query(sql, [newStatus, id, userId]);
 
     return {
         message: "Status updated successfully",
     };
 }
 
-async function deleteApplication(id) {
-
+async function deleteApplication(id, userId) {
     const sql = `
         DELETE FROM applications
         WHERE id = $1
+        AND user_id = $2
     `;
 
-    await pool.query(sql, [id]);
+    await pool.query(sql, [id, userId]);
 
     return {
         message: "Application deleted successfully",
